Moving back to st due to alacritty / opengl bug
I raised this as an alacritty bug but it looks like it might actually be
an issue with opengl itself? Not sure...not really wanting to open that
can of worms so st it is. Details of my bug reports are here:
    https://github.com/alacritty/alacritty/issues/4836

TL;DR - for whatever reason, alacritty doesn't handle WM_DELETE_WINDOW
correctly and results in zombie processes sitting around holding on to X
resource IDs which eventually leads to an exhaustion of available IDs.
Not fun...