Getting a systray working
Ideally I'd like to have a systray built into the statusbar itself but
for now this is working pretty well. I've moved back to using the root
window name widget so that the clock updates on a regular interval even
when window focus is not changing (need to sort the behaviour of the
built in widgets to avoid this issue at some point). I'm now using
'trayer' as a systray and I've tweaked things in the main penrose crate
to support a spacer widget which is able to conditionally render based
on the screen it is on.
Longer term I think I want the status bar as a whole to be configurable
per-screen (to help with differing resolutions) but that'll be more of a
change inside of penrose itself, potentially a breaking API change for
how status bar widgets are written.